requirements.txt
pytest
requests
faker
PyYAML
python-dotenv
pytest-html
allure-pytest
pytest-xdist
pika
prometheus-client
docker
